requirements.txt
torch>=1.0
pytorch-ignite
tqdm
click
numpy
torchvision
